[Gold V] Title: 최소비용 구하기, Time: 548 ms, Memory: 50148 KB -BaekjoonHub

+++ 백준/Gold/1916. 최소비용 구하기/최소비용 구하기.java
@@ -46,7 +46,7 @@ class Load implements Comparable<Load> {
 
     @Override
     public int compareTo(Load o) {
-        if (this.cost == o.cost) return (int) (this.idx - o.idx);
+        if (this.cost == o.cost) return this.idx - o.idx;
         return (int) (this.cost - o.cost);
     }
 }
